OJS 3.5 Compatibility
