+++ RosettaInfoSender.php
