+++ classes/xml/mods/ModsDOM.php
@@ -1,20 +1,20 @@
 <?php
 
 /**
- * @file plugins/importexport/rosetta/RosettaExportPlugin.php
+ * @file plugins/importexport/rosetta/classes/xml/mods/ModsDOM.php
  *
  * Copyright (c) 2014-2025 Simon Fraser University
  * Copyright (c) 2003-2025 John Willinsky
  * Distributed under the GNU GPL v3. For full terms see the file LICENSE.
  *
- * @class RosettaExportPlugin
+ * @class ModsDOM
  *
  * @ingroup plugins_importexport_rosetta
  *
  * @brief Rosetta export plugin
  */
 
-namespace APP\plugins\importexport\rosetta\classes\Mods;
+namespace APP\plugins\importexport\rosetta\classes\xml\mods;
 
 use APP\publication\Publication;
 use DOMDocument;
