--- index.php
@@ -1,13 +1,0 @@
-<?php
-
-/**
- * @file plugins/importexport/rosetta/index.php
- *
- * Copyright (c) 2014-2025 Simon Fraser University
- * Copyright (c) 2003-2025 John Willinsky
- * Distributed under the GNU GPL v3. For full terms see the file LICENSE.
- *
- * @brief Wrapper for rosetta export plugin.
- */
-
-return new \APP\plugins\importexport\rosetta\RosettaExportPlugin();
