+++ tests/classes/TestJournal.php
@@ -1,8 +1,24 @@
 <?php
 
-import('classes.article.ArticleGalley');
-import('classes.article.ArticleGalleyDAO');
-import('classes.journal.Journal');
+/**
+ * @file plugins/importexport/rosetta/tests/classes/TestIssue.php
+ *
+ * Copyright (c) 2014-2025 Simon Fraser University
+ * Copyright (c) 2003-2025 John Willinsky
+ * Distributed under the GNU GPL v3. For full terms see the file LICENSE.
+ *
+ * @class TestIssue
+ *
+ * @ingroup plugins_importexport_rosetta
+ *
+ * @brief Rosetta export plugin
+ */
+
+namespace APP\plugins\importexport\rosetta\tests\classes;
+
+use APP\journal\Journal;
+use APP\oai\ojs\OAIDAO;
+use PKP\db\DAORegistry;
 
 class TestJournal extends Journal
 {
@@ -52,7 +68,6 @@ class TestJournal extends Journal
 
     public function createOAI(): void
     {
-        import('classes.oai.ojs.OAIDAO');
         $oaiDao = $this->functionalRosettaExportTest->getMockBuilder(OAIDAO::class)
             ->setMethods(array('getJournal', 'getSection', 'getIssue'))
             ->getMock();
