+++ tests/classes/TestSubmission.php
@@ -1,13 +1,26 @@
 <?php
+/**
+ * @file plugins/importexport/rosetta/tests/classes/TestIssue.php
+ *
+ * Copyright (c) 2014-2025 Simon Fraser University
+ * Copyright (c) 2003-2025 John Willinsky
+ * Distributed under the GNU GPL v3. For full terms see the file LICENSE.
+ *
+ * @class TestIssue
+ *
+ * @ingroup plugins_importexport_rosetta
+ *
+ * @brief Rosetta export plugin
+ */
 
-import('classes.submission.Submission');
-import('plugins.importexport.rosetta.tests.classes.TestSection');
-import('plugins.importexport.rosetta.tests.classes.TestIssue');
-import('plugins.importexport.rosetta.tests.classes.TestPublication');
+namespace APP\plugins\importexport\rosetta\tests\classes;
+
+use APP\submission\Submission;
+use PKP\db\DAORegistry;
 
 class TestSubmission extends Submission
 {
-    public $locale;
+    public string $locale;
 
     public function __construct()
     {
