+++ tests/functional/RosettaFunctionsTest.php
@@ -1,26 +1,28 @@
 <?php
-declare(strict_types=1);
 
-require_mock_env('env2');
+/**
+ * @file plugins/importexport/rosetta/tests/functional/RosettaFunctionsTest.php
+ *
+ * Copyright (c) 2014-2025 Simon Fraser University
+ * Copyright (c) 2003-2025 John Willinsky
+ * Distributed under the GNU GPL v3. For full terms see the file LICENSE.
+ *
+ * @class RosettaFunctionsTest
+ *
+ * @ingroup plugins_importexport_rosetta
+ *
+ * @brief Rosetta export plugin
+ */
 
-import('plugins.importexport.rosetta.tests.functional.xml.mets.Mets');
-import('plugins.importexport.rosetta.tests.functional.xml.dublincore.DublinCore');
-import('plugins.importexport.rosetta.tests.classes.TestSubmission');
-import('plugins.importexport.rosetta.tests.classes.TestJournal');
+namespace APP\plugins\importexport\rosetta\tests\functional;
 
-import('plugins.importexport.rosetta.RosettaExportPlugin');
-import('plugins.importexport.rosetta.RosettaExportDeployment');
-import('lib.pkp.tests.plugins.PluginTestCase');
-import('lib.pkp.tests.plugins.metadata.MetadataPluginTestCase');
-
-import('lib.pkp.tests.PKPTestCase');
-
-import('lib.pkp.classes.oai.OAIStruct');
-import('lib.pkp.classes.oai.OAIUtils');
-import('plugins.oaiMetadataFormats.dc.OAIMetadataFormat_DC');
-import('plugins.oaiMetadataFormats.dc.OAIMetadataFormatPlugin_DC');
-import('lib.pkp.classes.core.PKPRouter');
-import('lib.pkp.classes.services.PKPSchemaService');
+use APP\core\Application;
+use APP\plugins\importexport\rosetta\tests\functional\xml\dublincore\DublinCore;
+use APP\plugins\importexport\rosetta\tests\functional\xml\mets\Mets;
+use PKP\core\PKPRouter;
+use PKP\plugins\Plugin;
+use PKP\plugins\PluginRegistry;
+use PKP\tests\plugins\PluginTestCase;
 
 class RosettaFunctionsTest extends PluginTestCase
 {
